+++ product_supplier_intercompany/models/__init__.py
@@ -1,3 +1,7 @@
-from . import pricelist
-from . import product
+from . import product_intercompany_supplier_mixin
+from . import product_pricelist_item
+from . import product_pricelist
+from . import product_template
+from . import product_product
+from . import product_supplierinfo
 from . import purchase_order
